requirements.txt
tensorflow==2.16.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
yfinance==0.2.18
pandas-ta==0.3.14b0
matplotlib==3.7.2
plotly==5.15.0
streamlit==1.37.0
flask==2.3.2
requests==2.31.0
python-telegram-bot==20.4
schedule==1.2.0
python-dotenv==1.0.0
seaborn==0.12.2
joblib==1.3.2
psutil==5.9.5

# Additional dependencies for LSTM pattern recognition
opencv-python-headless==4.8.0.76
tqdm==4.66.1
